+++ CHANGELOG.md
@@ -2,6 +2,19 @@ Release History
 ---------------
 ## [Unreleased]
 
+# [2.0.0] - 2017-03-28
+
+### Breaking change
+- `context` input must be declared like this `[context]="this"` in order for exemplify to be able to get markup through annotations (need for angular 4+ and typescript 2.2.0+).
+- `nested` input has been removed as it's no longer needed.
+- Link class names have been changed from `link-` to `exemplify-` ie. `link-copy` is now called `exemplify-copy`.
+
+### Added
+- Class to indicate if toggle code example is active or not, if active `exemplify-visible` will be added to link (see issue #5).
+
+### Fixed
+- Added `ngModel` to directives to be escaped (see issue #4)
+
 # [1.2.0] - 2017-03-09
 
 ### Added